Changed how the input is given
Running main.py with no arguments would scrape an empty domain name unless the domain variable was set in the code. Now, if no arguments are given, the usage statement is printed, instructing the user to supply a domain in the command-line. I figured the user could change the domain variable in the code if they didn't want to supply command-line input